initial implementation
This commit adds the initial implementation for the community
labeler action. It's a simple action that will add a label to an issue
or pull request when the creator is not a member of the configured
organization.

It will most probably change based on reviews and suggestions
from the team, however this is an MVP.