(MAINT) Add a guard clause to stop relabelling
Prior to this commit in some limited cases, it would be possible to
trigger the action on an issue or PR that already has the specified
label assigned. This is messy and would cause unnecessary notifications
for watchers and contributors of the repository.

This commit adds hasLabel method to check whether the current issue
already has the specified label applied. If true the action will return
gracefully. If false the action will continue to label the issue or PR.