(MAINT) Use a custom auth token
Prior to this commit, the test action was using secrets.GITHUB_TOKEN.
This token does not have the required permission to view org
membership,therefore the condition would always return false, creating
a label.

This commit updates the test action to use the CUSTOM_TOKEN and also
adds a note to the README regarding the required configuration.